# Conflicts:
#	externals/RoomClipper/src/RC_VicusClipping.h

+++ SIM-VICUS/src/SVPropBuildingSurfaceHeatingWidget.cpp
@@ -446,7 +446,7 @@ void SVPropBuildingSurfaceHeatingWidget::on_pushButtonSwitchControlZone_clicked(
 	QModelIndexList list = m_ui->tableWidgetSurfaceHeating->selectionModel()->selectedRows(4);
 	std::vector<VICUS::ComponentInstance> cis = SVProjectHandler::instance().project().m_componentInstances;
 
-	for(unsigned int i=0; i<list.size(); ++i) {
+	for(unsigned int i=0; i < (unsigned int)list.size(); ++i) {
 
 		int currentRow = list[i].row();
 		unsigned int idCi = m_ui->tableWidgetSurfaceHeating->item(currentRow, 4)->data(Qt::UserRole).toUInt();
